Add tmpdir workspace workaround fix for all user facing commands
