+++ bin/commands/components/install/cli.ts
@@ -12,21 +12,9 @@
 import * as index from './index';
 import * as std from 'std';
 import * as os from 'os';
-
-declare namespace console {
-  function log(...args:string[]): void;
-};
+import * as configmgr from '../../../libs/configmgr';
 
 index.execute(std.getenv('ZWE_CLI_PARAMETER_COMPONENT_FILE'), std.getenv('ZWE_CLI_PARAMETER_AUTO_ENCODING'), (std.getenv('ZWE_CLI_PARAMETER_SKIP_ENABLE') === 'true'));
 
-const tmpDir = std.getenv('ZWE_PRIVATE_TMP_MERGED_YAML_DIR');
-if (!std.getenv('PATH')) {
-  std.setenv('PATH','/bin:.:/usr/bin');
-}
-const rc = os.exec(['rm', '-rf', tmpDir],
-                       {block: true, usePath: true});
-if (rc == 0) {
-  console.log(`Temporary directory ${tmpDir} removed successfully.`);
-} else {
-  console.log(`Error: Temporary directory ${tmpDir} was not removed successfully, manual cleanup is needed. rc=${rc}`);
-}
+configmgr.cleanupTempDir();
+
