+++ bin/commands/stop/cli.ts
